+++ hslua-marshalling/test/HsLua/MarshallingTests.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE LambdaCase        #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-|
 Module      : HsLua.MarshallingTests
 Copyright   : © 2020-2021 Albert Krewinkel
@@ -10,7 +12,13 @@ Test marshalling of basic values.
 -}
 module HsLua.MarshallingTests (tests) where
 
+import Control.Monad ((<$!>))
+import HsLua.Core
+import HsLua.Marshalling.Peek
+import HsLua.Marshalling.Peekers
+import HsLua.Marshalling.Push
 import Test.Tasty (TestTree, testGroup)
+import Test.Tasty.HsLua ((=:), shouldBeResultOf)
 import qualified HsLua.Marshalling.PeekTests
 import qualified HsLua.Marshalling.PeekersTests
 import qualified HsLua.Marshalling.PushTests
@@ -23,4 +31,29 @@ tests = testGroup "Marshalling"
   , HsLua.Marshalling.PeekersTests.tests
   , HsLua.Marshalling.PushTests.tests
   , HsLua.Marshalling.UserdataTests.tests
+  , testGroup "nested"
+    [ "deeply nested list" =:
+      Success (mkDeeplyNested 33) `shouldBeResultOf` do
+        pushNested (mkDeeplyNested 33)
+        runPeek $ peekNested top
+    ]
   ]
+
+mkDeeplyNested :: Int -> Nested
+mkDeeplyNested i = foldr (\_ n -> List [n]) (Element i) [1..i]
+
+pushNested :: LuaError e => Pusher e Nested
+pushNested = \case
+  Element i   -> pushIntegral i
+  List nested -> pushList pushNested nested
+
+peekNested :: LuaError e => Peeker e Nested
+peekNested idx = liftLua (checkstack 2) >>= \case
+  False -> failPeek "stack overflow"
+  True -> liftLua (ltype idx) >>= \case
+    TypeNumber  -> Element <$!> peekIntegral idx
+    TypeTable   -> (List   <$!> peekList peekNested idx) `lastly` pop 1
+    _           -> failPeek "you dun goofed"
+
+data Nested = Element Int | List [Nested]
+  deriving (Eq, Show)